requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
redis>=5.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0
aiofiles>=23.2.0
discord-webhook>=1.3.0
croniter>=2.0.0
pyyaml>=6.0.0
rich>=13.0.0
typer>=0.9.0
psutil>=5.9.0
